Added AgentHosts asset to resolve hosts between AgentConfig and InstallConfig
